Fixed test and moved test files to their own folder
