+++ src/Workleap.DomainEventPropagation.Subscription.Tests/OfficevibeMigrationTests/SubscriptionOfficevibeMigrationTests.cs
@@ -1,12 +1,10 @@
 ﻿using System.Text.Json;
 using Azure.Messaging.EventGrid;
-using FakeItEasy;
-using GSoft.Extensions.DependencyInjection;
 using GSoft.Extensions.Xunit;
 using Microsoft.Extensions.DependencyInjection;
 using Officevibe.DomainEvents;
 
-namespace Workleap.DomainEventPropagation.Subscription.Tests;
+namespace Workleap.DomainEventPropagation.Subscription.Tests.OfficevibeMigrationTests;
 
 public sealed class SubscriptionOfficevibeMigrationTests : BaseIntegrationTest<OfficevibeSubscriptionMigrationFixture>
 {
@@ -22,10 +20,11 @@ public sealed class SubscriptionOfficevibeMigrationTests : BaseIntegrationTest<O
         var eventGridEvent = new EventGridEvent("subject", domainEvent.GetType().FullName, "1.0", BinaryData.FromString(serializedEvent));
 
         var webhookHandler = this.Services.GetRequiredService<IDomainEventGridWebhookHandler>();
-        var eventHandler = this.Services.GetRequiredService<IDomainEventHandler<OfficevibeEvent>>();
         await webhookHandler.HandleEventGridWebhookEventAsync(eventGridEvent, CancellationToken.None);
 
-        A.CallTo(() => eventHandler.HandleDomainEventAsync(A<OfficevibeEvent>._, A<CancellationToken>._)).MustHaveHappenedOnceExactly();
+        var testState = this.Services.GetRequiredService<SubscriberMigrationTestState>();
+
+        Assert.Equal(1, testState.OfficevibeDomainEventHandlerCallCount);
     }
 }
 
@@ -38,11 +37,13 @@ public sealed class OfficevibeSubscriptionMigrationFixture : BaseIntegrationFixt
         services.AddEventPropagationSubscriber()
             .AddDomainEventHandler<OfficevibeEvent, OfficevibeDomainEventHandler>();
 
-        services.DecorateWithSameLifetime<IDomainEventHandler<OfficevibeEvent>>(handler =>
-        {
-            return A.Fake<IDomainEventHandler<OfficevibeEvent>>(x => x.Wrapping(handler));
-        });
+        services.AddSingleton<SubscriberMigrationTestState>();
 
         return services;
     }
+}
+
+public class SubscriberMigrationTestState
+{
+    public int OfficevibeDomainEventHandlerCallCount { get; set; }
 }
